+++ lib/sidekiq/history.rb
@@ -6,8 +6,8 @@ end
 
 require 'sidekiq/api'
 require 'sidekiq/history/middleware'
-require 'sidekiq/history/statistic'
 require 'sidekiq/history/web_extension'
+require 'sidekiq/history/worker_statistic'
 require 'sidekiq/history/version'
 
 module Sidekiq