+++ lib/sidekiq/history/web_extension.rb
@@ -8,6 +8,8 @@ module Sidekiq
         end
 
         app.get '/history' do
+          @worker_statistic = Sidekiq::History::WorkerStatistic.new(20)
+          @tooltip_template = '<%= datasetLabel %> - <%= value %>'
           render(:erb, File.read(File.join(view_path, 'history.erb')))
         end
       end